% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{is_iv}
\alias{is_iv}
\title{Is \code{x} an iv?}
\usage{
is_iv(x)
}
\arguments{
\item{x}{\verb{[object]}

An object.}
}
\value{
A single \code{TRUE} or \code{FALSE}.
}
\description{
\code{is_iv()} tests if \code{x} is an iv object.
}
\examples{
is_iv(1)
is_iv(new_iv(1, 2))
}
