% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{iv-accessors}
\alias{iv-accessors}
\alias{iv_start}
\alias{iv_end}
\title{Access the start or end of an interval vector}
\usage{
iv_start(x)

iv_end(x)
}
\arguments{
\item{x}{\verb{[iv]}

An interval vector.}
}
\value{
The start or end of \code{x}.
}
\description{
\itemize{
\item \code{iv_start()} accesses the start of an interval vector.
\item \code{iv_end()} accesses the end of an interval vector.
}
}
\examples{
x <- new_iv(1, 2)

iv_start(x)
iv_end(x)
}
