% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv-package.R
\docType{package}
\name{iv-package}
\alias{iv}
\alias{iv-package}
\title{iv: Interval Types and Tools}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/iv}
  \item Report bugs at \url{https://github.com/DavisVaughan/iv/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
