% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs-deprecated.R
\name{iv-set-pairwise-deprecated}
\alias{iv-set-pairwise-deprecated}
\alias{iv_pairwise_complement}
\alias{iv_pairwise_union}
\alias{iv_pairwise_intersect}
\title{Pairwise set operations}
\usage{
iv_pairwise_complement(x, y)

iv_pairwise_union(x, y)

iv_pairwise_intersect(x, y)
}
\arguments{
\item{x, y}{\verb{[iv]}

A pair of interval vectors.

These will be cast to the same type, and recycled against each other.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated in favor of their \code{set_} prefixed equivalents.
\itemize{
\item \code{iv_pairwise_complement()} -> \code{\link[=iv_pairwise_set_complement]{iv_pairwise_set_complement()}}
\item \code{iv_pairwise_union()} -> \code{\link[=iv_pairwise_set_union]{iv_pairwise_set_union()}}
\item \code{iv_pairwise_span()} -> \code{\link[=iv_pairwise_set_span]{iv_pairwise_set_span()}}
\item \code{iv_pairwise_intersect()} -> \code{\link[=iv_pairwise_set_intersect]{iv_pairwise_set_intersect()}}
\item \code{iv_pairwise_difference()} -> \code{\link[=iv_pairwise_set_difference]{iv_pairwise_set_difference()}}
\item \code{iv_pairwise_symmetric_difference()} -> \code{\link[=iv_pairwise_set_symmetric_difference]{iv_pairwise_set_symmetric_difference()}}
}
}
\keyword{internal}
