% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{iv-sets}
\alias{iv-sets}
\alias{iv_set_complement}
\alias{iv_set_union}
\alias{iv_set_intersect}
\title{Set theoretical operations}
\usage{
iv_set_complement(x, ..., lower = NULL, upper = NULL)

iv_set_union(x, y)

iv_set_intersect(x, y)
}
\arguments{
\item{x}{\verb{[iv]}

An interval vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{lower, upper}{\verb{[vector(1) / NULL]}

Bounds for the universe over which to compute the complement. These should
have the same type as the element type of the interval vector. It is
often useful to expand the universe to, say, \code{-Inf} to \code{Inf}.}

\item{y}{\verb{[iv]}

An interval vector.}
}
\value{
For \code{iv_set_complement()}, a vector of the same type as \code{x} containing the
complement.

For all other set operations, a vector of the same type as the common type of
\code{x} and \code{y} containing the result.
}
\description{
This family of functions treats ivs as sets. They always compute
the \link[=iv_union]{minimal} iv of each input and return a minimal iv.
\itemize{
\item \code{iv_set_complement()} takes the complement of a set of intervals. By
default, the minimum and maximum of the inputs define the bounds to take
the complement over, but this can be adjusted with \code{lower} and \code{upper}.
Missing intervals are always dropped in the complement.
\item \code{iv_set_union()} takes the union of two sets of intervals. It is equivalent
to combining the two vectors together and then calling \code{iv_union()}.
\item \code{iv_set_intersect()} takes the intersection of two sets of intervals.
}
}
\examples{
x <- iv_pairs(
  c(10, 12),
  c(0, 5),
  c(NA, NA),
  c(3, 6),
  c(-5, -2),
  c(NA, NA)
)
x

y <- iv_pairs(
  c(2, 7),
  c(NA, NA),
  c(-3, -1),
  c(14, 15)
)
y

# Complement contains any values from `[-5, 12)` that aren't represented
# in these intervals. Missing intervals are dropped.
iv_set_complement(x)

# Expand out the "universe" of possible values
iv_set_complement(x, lower = -Inf)
iv_set_complement(x, lower = -Inf, upper = Inf)

# Union returns the minimal iv containing all of the information from
# both interval vectors
iv_set_union(x, y)

# Minimal iv representing the set intersection
iv_set_intersect(x, y)
}
