% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{iv_set_complement}
\alias{iv_set_complement}
\title{Complement}
\usage{
iv_set_complement(x, ..., lower = NULL, upper = NULL)
}
\arguments{
\item{x}{\verb{[iv]}

An interval vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{lower, upper}{\verb{[vector(1) / NULL]}

Bounds for the universe over which to compute the complement. These should
have the same type as the element type of the interval vector. It is
often useful to expand the universe to, say, \code{-Inf} to \code{Inf}.}
}
\value{
An object with the same type as \code{x} representing its complement.
}
\description{
\code{iv_set_complement()} computes the set theoretical complement of an interval
vector. The "universe" to compute the complement over is considered to be
the minimal and maximal values of \code{x}, but this can be adjusted with \code{lower}
and \code{upper}.
}
\examples{
x <- new_iv(c(10, 0, 3, -5), c(12, 5, 6, -2))
x

# Complement contains any values from `[-5, 12)` that aren't represented
# in these intervals
iv_set_complement(x)

# Expand out the "universe" of possible values
iv_set_complement(x, lower = -Inf)
iv_set_complement(x, lower = -Inf, upper = Inf)
}
