% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv.R
\name{new_iv}
\alias{new_iv}
\title{Construct a new iv}
\usage{
new_iv(start, end, ..., class = character())
}
\arguments{
\item{start, end}{\verb{[vector]}

A pair of vectors to represent the bounds of the intervals.

To be a valid interval vector, \code{start} must be strictly less than \code{end},
or both \code{start} and \code{end} must be a missing value.}

\item{...}{\verb{[name-value pairs]}

Additional named attributes to attach to the result.}

\item{class}{\verb{[character]}

The name of the subclass to create.}
}
\description{
\code{new_iv()} is a developer focused function for creating a new interval
vector. It does minimal checks on the inputs, for performance.
}
\examples{
new_iv(1, 2)
}
