% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv-package.R
\docType{package}
\name{iv-package}
\alias{iv-package}
\alias{_PACKAGE}
\title{iv: Interval Vectors}
\description{
Provides a library for generic interval manipulations using a new interval vector class. Capabilities include: locating various kinds of relationships between two interval vectors, merging overlaps within a single interval vector, splitting an interval vector on its overlapping endpoints, and applying set theoretical operations on interval vectors. Many of the operations in this package were inspired by James Allen's paper, Maintaining Knowledge about Temporal Intervals <https://cse.unl.edu/~choueiry/Documents/Allen-CACM1983.pdf>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/iv}
  \item \url{https://davisvaughan.github.io/iv/}
  \item Report bugs at \url{https://github.com/DavisVaughan/iv/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
