% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs-deprecated.R
\name{iv-sets-deprecated}
\alias{iv-sets-deprecated}
\alias{iv_complement}
\alias{iv_union}
\alias{iv_intersect}
\alias{iv_difference}
\alias{iv_symmetric_difference}
\title{Set operations}
\usage{
iv_complement(x, ..., lower = NULL, upper = NULL)

iv_union(x, y)

iv_intersect(x, y)

iv_difference(x, y)

iv_symmetric_difference(x, y)
}
\arguments{
\item{x}{\verb{[iv]}

An interval vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{lower, upper}{\verb{[vector(1) / NULL]}

Bounds for the universe over which to compute the complement. These should
have the same type as the element type of the interval vector. It is
often useful to expand the universe to, say, \code{-Inf} to \code{Inf}.}

\item{y}{\verb{[iv]}

An interval vector.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated in favor of their \code{set_} prefixed equivalents.
\itemize{
\item \code{iv_complement()} -> \code{\link[=iv_set_complement]{iv_set_complement()}}
\item \code{iv_union()} -> \code{\link[=iv_set_union]{iv_set_union()}}
\item \code{iv_intersect()} -> \code{\link[=iv_set_intersect]{iv_set_intersect()}}
\item \code{iv_difference()} -> \code{\link[=iv_set_difference]{iv_set_difference()}}
\item \code{iv_symmetric_difference()} -> \code{\link[=iv_set_symmetric_difference]{iv_set_symmetric_difference()}}
}
}
\keyword{internal}
